[FIX] web_kanban: fix problems introduced by someone
Someone really smart decided to change web_kanban by moving something
from web to web_kanban.  This makes sense, but it is a problem because
the code in web is not shared between enterprise and community, but
web_kanban is shared.

Also, it was put in kanban_view.  If we have a module system in
javascript, is might be a good idea to use it.

So, this commit moves the code in a new file, and update the code to
the last version from enterprise. Yay for two codebase in parallel.