[FIX] web_gantt: render gantt diagram in DOM
Action and view managers render theirs elements in document fragments, and
append them to the DOM once they are ready. The dhtmlxgantt v.1.3 requires that
the gantt diagram in rendered in the DOM. This rev. is an hack to ensure that
this is always the case.

Note that this is no more a requirement with dhtmlxgantt v.3.1.