[FIX] account_followup: contact with type = "invoice"
In "Payment Follow-up" when clicking on "Send Overdue Email"
if contact with type = "invoice"  and an email exists in the childs then
the email will be send to this contact else the normal behavior
is kept. When the overvue email is sent to a contact who is not the commercial
partner, a message is written in the chatter of the commercial partner to inform him
that the email has been sent to the "invoice" contact.

Closes #7870
opw:646149, 646575