[FIX] website_sale: force refresh cart without cache when product removed
Firefox keep in cache the value of input.
So, when we delete a product line and call location.reload(), firefox
remember the value and state of this input and don't refresh with the new
value.

Adding True as param (forceReload) fixes the problem since it doesn't
use old value. Another way will be to add autocomplete="off".

ForcedReload (Optional): Is a Boolean flag, which, when it is true, causes the
page to always be reloaded from the server. If it is false or not specified,
the browser may reload the page from its cache.

This closes #7888 #3342 #7491