[FIX] stock: no uom rounding for DO availability & reservation
Instead of rounding the availability and reversation
of a delivery order line to the current UOM rounding,
we round to the generic Product of measure precision.

This is to have the real availability, even for a unit
of measure that normally cannot be split.

For instance, for a product for which the default
uom is 'Unit(s)', but for a specific delivery order
you deliver 1 'Dozens' of this product,
if you have 6 Units available, it will now display
that 0.500 'Dozens' is available and reserved, which
is what is actually happening.

Before, it displayed that 1 dozens was available, even
if only 6 units were available, on the 12 needed.

opw-643312