[FIX] survey: partner access to survey input
Allow the partner associated to a survey user input to read the data on
this output.

A survey can be created by a different user than the one
filling it, for example an appraisal interview request can be created by
another user than its interviewer. In this case if the interviewer is
not Survey / Manager an error would happen.

closes #7978
opw-644791