[FIX] crm: prevent use a sales tag from another team.
On the `categ_ids` fields definition on `crm.lead`,
a domain has been added in order to prevent
the use of a sales tag belonging to another
sales team.

As the domain was overidden in the form views,
with a different domain, it was allowed
to use a sales tag from another sales team.

Besides, the domain applied in the form views,
`[('object_id.model','=','crm.lead')]`
is included in the domain applied in the fiel
definition.

Closes #2898