[FIX] web,base: list view: limit
Before this rev., the number of records per page to display in list views
could be specified in the action definition, but it didn't work for One2Many
list views (because they are instanciated by a FormView, which isn't supposed
to have a reference to the action, instead of the ActionManager as regular
list views).

This rev. allows to specify the list_view's limit directly in the definition
of the view, and not only on the action as before.