[FIX] stock: barcode print in stock reports.
During 32fcfa6, `ean13` was massively renamed to `barcode`.

Nevertheless, the barcode controller (/report/barcode)
doesn't accept `barcode` as a valid barcode type, it's
actually `ean13`.

In the report printing barcodes,
(in report_stockpicking.xml)
this `ean13` shouldn't
have been renamed to `barcode`.

The `barcode` in report_stockpicking_operations.xml
is not due to this revision, but is probably the
result of a copy/paste.

opw-647263