[FIX] purchase_analytic_plans: apply purchase analytic group on analytic account.
When the Purchase analytic plans module is not installed,
the analytic account field in the lines of the purchase order
is available only for the group
`Analytic Accounting for Purchases`

It should be the same for the field of the analytic distribution plan.

Otherwise, it will lead to an access error if you don't, as
you do not have access to the required model.

opw-647606