[FIX] sale: in a SO line, changing uom qty must change uos qty
This is a oversight of the forward port
af07b2a07587d83a67928cf30dd9172f2c870cb7
from 8.0

The revision in 8.0 is
9a9720ac14d3508e93339e351ba19de4b3ca2fd6

Not updating the product uos qty on updating the uom qty
is wrong, and it has multiple side-effect.

The SO line margin not being correctly recomputed
when changing the quantity is one of them.

opw-647810