[IMP] stock: picking validate button + onchange uom
The button Mark As Todo should be before the button Validate because
most people check things when they transfer stuff in their warehouse.
Moreover, it creates confusion, because people will think Validate means Confirm
(or Mark As Todo).

When validating an empty picking, it will give an error now.

On_change of product on stock moves and pack operations should return
a domain for the UoM.