[IMP] point_of_sale: reimplement cash control support
Support for cash control was removed in
c04065abd8f62c9a211c8fa824f5eecf68e61b73. Because the cash_control field
on account.journal was removed, we now define a cash_control preference
on pos.config. In order to set the opening and closing balances the new
AccountBankStmtCashWizard is used.