[IMP] models: optimize recomputations by saving values in record batches
The recomputation of a field `name` on `recs` is done roughly like this:

    for rec in recs:
	rec._write({name: rec[name]})

The effect of this patch is to group records by value `rec[name]`, and to
update records in batches like:

    recs.write({name: val})             # rec[name] == val for all rec in recs

As a consequence, less queries are issued to the database for updating records.
This improvement is very effective in the case of migrations!