[IMP] sale_expense, hr_expense: test more robust
The test does not have hardcoded values referring to demo data anymore.
Moreover, in the test, we now use values which are different from the
demo data (621.54 instead of 700). Overall, this will make the test more
robust.

hr_expense: change car travel sale price
Car Travel Expense is used as a product invoiced at sale price.
Therefore, it is more useful to have a cost price and a sale price which
are different.

The whole idea of this commit is to avoid having the same values
everywhere.