[FIX] sale: multiple fixes
1) Check unit price before linking SO line

We need to make sure that the unit price on the SO line is the same than
the unit price that will be applied for the new analytic line.

It is the case for an expense product which is invoiced at cost price to
the customer, for example a flight ticket. The user should be able to
link 2 flight tickets with different prices on the same SO, which is not
possible before this fix.

2) Convert account move line amount into currency of SO if necessary

3) Partial revert of 2d5776d248920267bcf6d002e4420dd3d6a8282e

We want to decrease the invoiced quantity if the refund is generated
from the SO, but not if the refund is generated from the invoice
directly. Indeed, if we decrease the quantity from a refund generated
from the invoice, this will make the SO line invoiceable again, which is
something we don't want.

Fortunately for us, a refund generated from an invoice is not linked to
the SO (see method _refund_cleanup_lines in account_invoice.py), so no
extra code is needed.

4) Iterate only once in the SO lines in order to find out the
invoice_status of the SO