[FIX] account_cancel: allow cancel of whole bank statement
This revision is related to 6f355623f0632527aae55d02d24a5a06976441c2.

The above revision had as goal to not allow the partial cancellation
(line per line) of a confirmed bank statement, as it would
prevent the ability to reconcile the statement, as reconcilling
a confirmed statement it not possible.

Nevertheless, cancelling the whole bank statement should be allowed
(using the `Cancel Statement` button), as this cancel all lines
and re-set the statement in draft.

opw-649864
fixes #8568