[IMP] res_partner : name field is now optional for addresses.
The name_get has been updated accordingly : if an address has no name, we now show parent_name, res_partner.type.
The contact template shows the name when it exists, and the parent_name if it does not. This should produce cleaner reports.

Though the name field is no longer required on the res.partner model,
"required" attribute in views and an sql contraint enforce that all contacts have a name.