[FIX] point_of_sale: allow opening of sessions with cash control enabled
Without cash control the pos_session workflow automatically sets session
states to 'open' (and so they skip 'opening_control'). With cash control
enabled however this does not happen.
8a232e7f45eb52f6e5589ff1183d97bbd354cb7e removed a button which allowed
you to open the session. This adds back a simplified button which
appears when necessary and just sets the state to 'open'.