[FIX] ir_qweb: useless test in js bundle generation
If one of the javascript files of the bundle has been modified since the
bundle generation, we are sure that get_attachments will return an empty
list because it searches with the version of the bundle which takes into
account the last modification of any of the bundle's content.

Also, this check is not present in the css bundle generation.